Add endpoint, loading and success states
