+++ package.json
@@ -48,6 +48,7 @@
     "react-dom": "^16.14.0",
     "react-fast-marquee": "^1.2.1",
     "react-helmet": "^6.1.0",
+    "react-mailchimp-subscribe": "^2.1.3",
     "react-markdown": "^8.0.3",
     "react-moment": "^1.1.1",
     "react-paginate": "^7.1.3",
